split apart tests, refactor into fixtures, make all pass
